[refactor] fetchPlayersDataを分離
