+++ spa/services/tournament/WsTournamentMatchingManager.js
@@ -1,10 +1,10 @@
-import fetchApiNoBody from "../../api/fetchApiNoBody.js";
+import fetchPlayersData from "../../api/fetchPlayersData.js";
 import { renderMatchingRoom } from "../../components/MatchingRoom.js";
 import { renderWaitOrStart } from "../../components/WaitOrStart.js";
 import config from "../../config.js";
 import SPA from "../../spa.js";
 import stateManager from "../../stateManager.js";
-import { renderMatchingInfo } from "./MatchingInfo.js";
+import { renderMatchingInfo } from "./TournamentMatchingInfo.js";
 
 const wsEventHandler = {
   handleOpen(message) {
@@ -47,22 +47,6 @@ const wsEventHandler = {
     console.error("WebSocket error:", message);
   },
 };
-
-async function fetchPlayersData(userIdList) {
-  try {
-    const promises = userIdList.map((id) =>
-      fetchApiNoBody("GET", config.userService, `/users?userid=${id}`),
-    );
-    const results = await Promise.all(promises);
-    const playersData = results.map((item) => ({
-      avatarPath: `${config.userService}/${item.data.avatarPath}`,
-      name: item.data.username,
-    }));
-    return playersData;
-  } catch (error) {
-    console.error("Error");
-  }
-}
 
 const WsTournamentMatchingManager = {
   socket: null,
